# Optimize `min_boolean` and `bool_and`